Combined ABC for stack and queue and modified stack_queue sub package name. (#19)
