+++ data_structures/sequence/stack_queue/array_stack.py
@@ -6,7 +6,7 @@ and looks dumb, because it only serves as an data structure exercise and has no
 practical usage.
 """
 from typing import TypeVar, Optional, Sequence
-from .custom_stack import CustomStack
+from .custom_stack_queue import CustomStack
 from .size_mixin import SizeMixin
