+++ data_structures/sequence/stack_queue/size_mixin.py
