+++ data_structures/sequence/stack_queue/stacked_queue.py
@@ -6,7 +6,7 @@ dumb, because it only serves as an data structure exercise and has no practical
 usage.
 """
 from typing import TypeVar, Optional, Sequence
-from .custom_queue import CustomQueue
+from .custom_stack_queue import CustomQueue
 from .linked_stack import LinkedStack
 from .size_mixin import SizeMixin
